pyfuncperfsample/requirements.txt
azure-functions
pyinstrument
memory-profiler